+++ tests/mod.rs
@@ -7,6 +7,7 @@ mod core {
 }
 
 mod command {
+    mod test_amend;
     mod test_hide;
     mod test_init;
     mod test_move;